+++ business-center/business-center-service/src/main/java/com/gxa/jbgsw/business/service/BillboardService.java
@@ -36,7 +36,7 @@ public interface BillboardService extends IService<Billboard> {
 
     void updateMyBillboard(BillboardDTO billboardDTO);
 
-    int getMyReceiveBillboard(Long userId, Integer trueType);
+    long getMyReceiveBillboard(Long userId, Integer trueType);
 
     PageResult<MyReceiveBillboardInfo> queryMyReceiveBillboard(MyReceiveBillboardRequest request);
